Split GUI window chrome tests
